pyproject.toml
[tool.poetry]
name = "django-shorten-urls"
version = "0.1.0"
description = ""
authors = ["Samsul Hadi"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
django = "==4.2.11"
psycopg2-binary = "==2.9.9"
djangorestframework-simplejwt = "==5.3.1"
django-filter = "==23.3"
pillow = "==10.3.0"
gunicorn = "==21.2.0"
django-cors-headers = "==4.3.0"
django-solo = "==2.1.0"
django-prefix-id = "==1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3"
pytest-cov = ">=4.1.0"
pytest-django = ">=4.5.2"
flake8 = ">=6.1.0"
ipython = ">=8.16.1"
parameterized = ">=0.9.0"
black = ">=23.10.1"
isort = ">=5.12.0"
django-extensions = ">=3.2.3"
django-cors-headers = ">=4.3.0"
mixer = ">=7.2.2"
coverage = ">=7.3.2"
pylint = "==3.0.2"
pylint-django = "==2.5.4"
pylint-plugin-utils = "==0.8.2"
bandit = "==1.7.7"
liccheck = "==0.9.2"
safety = ">=2.3.4"
selenium = ">=4.14.0"
drf-spectacular = ">=0.27.2"
drf-spectacular-sidecar = ">=2024.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
